requirements.txt
asgiref==3.11.0
attrs==25.4.0
black==25.11.0
click==8.3.1
dj-database-url==3.0.1
django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pyjwt==2.10.1
python-decouple==3.8
pytokens==0.3.0
pyyaml==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.3
uritemplate==4.2.0
